+++ burst-kotlin-plugin/src/main/kotlin/app/cash/burst/kotlin/TestInterceptorIrGenerationExtension.kt
@@ -54,5 +54,7 @@ class TestInterceptorIrGenerationExtension(
     }
 
     moduleFragment.transform(transformer, null)
+
+    log(moduleFragment.dump())
   }
 }